moved all test cases to root dir
