+++ jest.config.ts
@@ -1,13 +1,13 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  rootDir: 'packages',
-  transform: {
-    '^.+\\.ts?$': 'ts-jest',
-  },
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.test.json'
-    }
-  }
-};
+	preset: "ts-jest",
+	testEnvironment: "node",
+	rootDir: "tests",
+	transform: {
+		"^.+\\.ts?$": "ts-jest",
+	},
+	globals: {
+		"ts-jest": {
+			tsconfig: "tsconfig.test.json",
+		},
+	},
+}
